Changed long to int, so it will be 32-bits for both MSVC and GCC 64-bits.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7806 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/odbc_xsb.c
@@ -1822,7 +1822,7 @@ int GetColumn(CTXTdecl)
     {
       Cell h;
       Integer intvalue;
-      intvalue = (Integer)(*(long *)(cur->Data[ColCurNum]));
+      intvalue = (Integer)(*(int *)(cur->Data[ColCurNum]));  // int is 32 bit in unix
       bld_oint(&h,intvalue);
       return unify(CTXTc op, h);
     }
